+++ src/index.ts
@@ -1,11 +1,13 @@
 import 'reflect-metadata';
 import PlatformTools from './PlatformTools';
 import MetaSchemaStorage from "./MetaSchemaStorage";
+import {ConvictModel} from './ConvictModel';
 
 export * from './decorators/Property';
 export * from './decorators/Config';
-export * from './ConvictModel';
-
+export {
+    ConvictModel
+};
 /**
  * Gets schema storage from the global
  */
@@ -16,6 +18,14 @@ export function getMetaSchemaStorage(): MetaSchemaStorage {
         globalScope.convictMetaSchemaStorage = new MetaSchemaStorage();
     }
     return globalScope.convictMetaSchemaStorage;
+}
+
+/**
+ * Quickly get a convict model
+ * @param entities
+ */
+export function getConvictModel(entities?: Array<((new ()=>{})|string)>): ConvictModel {
+    return new ConvictModel(entities);
 }
 
 export interface SchemaRepoItem {