+++ src/test/scenarios/simple_flat/SimpleFlat.model.ts
@@ -1,0 +1,12 @@
+import {Property} from "../../../";
+
+export default class SimpleFlat {
+
+    @Property({
+        doc: 'The name of the thing',
+        default: 'Convict',
+        env: 'MY_CONFIG_NAME'
+    })
+    public name: string;
+
+}