+++ src/test/scenarios/sub_no_main/SubNoMainChild.model.ts
@@ -1,0 +1,12 @@
+import {Property} from "../../../";
+
+export default class SubConfig {
+
+    @Property({
+        doc: 'The bar',
+        default: 0,
+        env: 'SUB_CONFIG_BAR'
+    })
+    public bar: number;
+
+}